% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_odm.R
\name{build_odm_args}
\alias{build_odm_args}
\title{Build an ODM Docker command}
\usage{
build_odm_args(
  dataset_dir,
  project_name = "micasense",
  image = "opendronemap/odm",
  radiometric_calibration = "camera+sun",
  orthophoto_resolution_cm = 5,
  max_concurrency = 4,
  fast_orthophoto = TRUE,
  build_dsm = FALSE,
  build_dtm = FALSE,
  pc_las = FALSE,
  pc_csv = FALSE,
  pc_copc = FALSE,
  tiles = FALSE,
  three_d_tiles = FALSE,
  gltf = FALSE,
  rerun_from = NULL,
  end_with = NULL,
  extra_args = character()
)
}
\arguments{
\item{dataset_dir}{Host folder mounted to \verb{/datasets}.}

\item{project_name}{ODM project name inside \code{dataset_dir}.}

\item{image}{Docker image name.}

\item{radiometric_calibration}{ODM radiometric calibration option.}

\item{orthophoto_resolution_cm}{Orthophoto resolution in centimeters.}

\item{max_concurrency}{Maximum concurrent ODM workers.}

\item{fast_orthophoto}{Logical. Add \code{--fast-orthophoto}.}

\item{build_dsm}{Logical. Add DSM generation.}

\item{build_dtm}{Logical. Add DTM generation options.}

\item{pc_las}{Logical. Export LAS point cloud.}

\item{pc_csv}{Logical. Export CSV point cloud.}

\item{pc_copc}{Logical. Export COPC point cloud.}

\item{tiles}{Logical. Export web map tiles.}

\item{three_d_tiles}{Logical. Export 3D tiles.}

\item{gltf}{Logical. Export glTF model.}

\item{rerun_from}{Optional ODM stage.}

\item{end_with}{Optional ODM end stage.}

\item{extra_args}{Additional ODM arguments.}
}
\value{
Character vector of arguments for \code{system2("docker", args)}.
}
\description{
Build an ODM Docker command
}
