% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{default_rgb_band_map}
\alias{default_rgb_band_map}
\title{Default band map for a 3-band RGB orthomosaic (Sony / DJI / Phantom)}
\usage{
default_rgb_band_map()
}
\value{
Named integer vector with \code{Red}, \code{Green}, \code{Blue}.
}
\description{
Default band map for a 3-band RGB orthomosaic (Sony / DJI / Phantom)
}
