% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_pointcloud.R
\name{points_in_roi}
\alias{points_in_roi}
\title{Test whether coordinates are inside a polygon ROI}
\usage{
points_in_roi(x, y, roi_polygon)
}
\arguments{
\item{x}{Numeric x coordinates.}

\item{y}{Numeric y coordinates.}

\item{roi_polygon}{Data frame with \code{x} and \code{y} polygon vertices.}
}
\value{
Logical vector.
}
\description{
Test whether coordinates are inside a polygon ROI
}
\examples{
roi <- data.frame(x = c(0, 5, 5, 0), y = c(0, 0, 5, 5))
points_in_roi(x = c(1, 6), y = c(1, 6), roi_polygon = roi)
}
