% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field.R
\name{read_field_data}
\alias{read_field_data}
\title{Read field biomass data}
\usage{
read_field_data(path)
}
\arguments{
\item{path}{CSV path.}
}
\value{
A data frame.
}
\description{
Read field biomass data
}
\examples{
field_path <- system.file("extdata", "field_samples.csv", package = "DroneBioR")
field <- read_field_data(field_path)
head(field)
}
