% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{summarize_spatraster}
\alias{summarize_spatraster}
\title{Summarize a SpatRaster by layer}
\usage{
summarize_spatraster(x, fun = c("min", "mean", "max", "sd"))
}
\arguments{
\item{x}{A \code{terra::SpatRaster}.}

\item{fun}{Summary functions supported by \code{terra::global()}.}
}
\value{
A data frame.
}
\description{
Summarize a SpatRaster by layer
}
