% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_webodm.R
\name{webodm_list_projects}
\alias{webodm_list_projects}
\title{List WebODM projects}
\usage{
webodm_list_projects(base_url, token)
}
\arguments{
\item{base_url}{Root URL of the WebODM server.}

\item{token}{JWT from \code{\link[=webodm_authenticate]{webodm_authenticate()}}.}
}
\value{
A data frame with \code{id}, \code{name}, \code{description}, \code{created_at}.
}
\description{
List WebODM projects
}
