% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_pointcloud.R
\name{build_chm_from_dsm_dtm}
\alias{build_chm_from_dsm_dtm}
\title{Build a canopy height model from DSM and DTM rasters}
\usage{
build_chm_from_dsm_dtm(dsm_path, dtm_path)
}
\arguments{
\item{dsm_path}{DSM GeoTIFF path.}

\item{dtm_path}{DTM GeoTIFF path.}
}
\value{
A \code{terra::SpatRaster} with non-negative canopy height in meters.
}
\description{
Build a canopy height model from DSM and DTM rasters
}
\examples{
dsm <- system.file("extdata", "dsm_subset.tif", package = "DroneBioR")
dtm <- system.file("extdata", "dtm_subset.tif", package = "DroneBioR")
chm <- build_chm_from_dsm_dtm(dsm, dtm)
terra::minmax(chm)
}
