% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field.R
\name{extract_field_spectral_data}
\alias{extract_field_spectral_data}
\title{Extract raster values at field sample points}
\usage{
extract_field_spectral_data(
  field_data,
  predictors,
  predictor_crs = terra::crs(predictors)
)
}
\arguments{
\item{field_data}{Field data frame.}

\item{predictors}{Raster stack with bands and indices.}

\item{predictor_crs}{CRS of x/y coordinates when x/y are used.}
}
\value{
A data frame with field columns and extracted raster values.
}
\description{
Extract raster values at field sample points
}
