% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_pointcloud.R
\name{filter_points_by_roi}
\alias{filter_points_by_roi}
\title{Filter point cloud data by a polygon ROI}
\usage{
filter_points_by_roi(points, roi_polygon)
}
\arguments{
\item{points}{Data frame with \code{x} and \code{y} coordinates.}

\item{roi_polygon}{Data frame with \code{x} and \code{y} polygon vertices.}
}
\value{
Filtered data frame.
}
\description{
Filter point cloud data by a polygon ROI
}
