% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/products.R
\name{odm_product_paths}
\alias{odm_product_paths}
\title{Return expected ODM product paths}
\usage{
odm_product_paths(project)
}
\arguments{
\item{project}{A \code{dronebio_project} object.}
}
\value{
Named character vector of expected product paths.
}
\description{
Return expected ODM product paths
}
