% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{run_drone_biomass_studio}
\alias{run_drone_biomass_studio}
\title{Start Drone Biomass Studio}
\usage{
run_drone_biomass_studio(
  project_dir = getwd(),
  port = NULL,
  launch.browser = TRUE,
  ...
)
}
\arguments{
\item{project_dir}{Default project directory.}

\item{port}{Optional local port.}

\item{launch.browser}{Logical. Open a browser window.}

\item{...}{Additional arguments passed to \code{shiny::runApp()}.}
}
\value{
The result of \code{shiny::runApp()}.
}
\description{
Start Drone Biomass Studio
}
\examples{
\dontrun{
run_drone_biomass_studio(project_dir = "/path/to/Drone_Biomass")
}
}
