% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{scale_to_reflectance}
\alias{scale_to_reflectance}
\title{Scale raster values to reflectance-like 0-1 values}
\usage{
scale_to_reflectance(x, scale_factor = NULL)
}
\arguments{
\item{x}{A \code{terra::SpatRaster}.}

\item{scale_factor}{Optional numeric scale factor.}
}
\value{
A \code{terra::SpatRaster}.
}
\description{
Scale raster values to reflectance-like 0-1 values
}
\examples{
ortho_path <- system.file("extdata", "micasense_subset.tif", package = "DroneBioR")
ortho <- read_multispectral_orthomosaic(ortho_path)
refl <- scale_to_reflectance(ortho$bands)
terra::minmax(refl)
}
