% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/products.R
\name{summarize_odm_products}
\alias{summarize_odm_products}
\title{Summarize available ODM products}
\usage{
summarize_odm_products(project)
}
\arguments{
\item{project}{A \code{dronebio_project} object.}
}
\value{
A data frame with product, path, availability and file size.
}
\description{
Summarize available ODM products
}
